Create a PermalinkFactory to avoid injecting the service in SDK classes
